requirement.txt
fastapi
uvicorn
streamlit
langchain
langchain-community
langchain-ollama
langchain-chroma
langchain-huggingface
pydantic
python-multipart
requests
beautifulsoup4
lxml
tiktoken
sentence-transformers
google-auth
google-auth-oauthlib
google-auth-httplib2
google-generativeai